update
added function select all from database except image extensions we are getting a list of data which they are not images. That means for example .txt file.  added check in files and directories if newfile.txt is available skip it if it is do not add to array of elemenst we need to make a check for directories too need a list of directories for scan, then we check value of $file if it is equal, then skip it do not add it to array. added function maintain database it will delete all the recrds from scanned data table which they are not image formats. added database backup. added call function in body template for index php for maintaining database.