update
added check if user is logged in do not show registration and login form if user is not login show it added check if there was post method and if there is any user who matched username and password in user json if there is login user add to session variable user and login in time in seconds. else print error and form for login in podnožje ve put a check to set status logged in or not logged in if there is session user put logged in in the table if not put not logged in into table added new functions for reading data from json and  for destoying session data and to return time how much session lasted. modified style for error and forma  added in zaglavlje check if session exists if session do not exists start new session