update
added in notepad data about admin and regular user username and password it is recommended to change it when this application is deployed on server.  added dcl script for dcl grant rights on some user.  added image galley fix which fix name of the table for foreign key.  added two users defined in text data in database readme into the database for user validation if user exists rights will be granted. need to write procedure for granting rights.